requirements.txt
torch
seaborn
pandas
numpy
munkres
torch-sparse
torch-scatter
torch-geometric
torch-cluster
PyYAML
